Make bot environment-specific
Now the bot can be a test bot or a live bot. The live bot will respond
to the commands normally. The test bot will only respond to commands
that start with "test". Also made commands not case-sensitive. Added
notification that bot started up, so I can know when the bot went down
for whatever reason.